[DCGEO-59] Update unit tests
