+++ apps/bfd-server/bfd-server-war/src/main/java/gov/cms/bfd/server/war/r4/providers/preadj/ClaimTypeV2.java
@@ -52,7 +52,7 @@ public enum ClaimTypeV2 implements ResourceTypeV2<Claim> {
   }
 
   /**
-   * @return the {@link ClaimTypeTransformerV2} to use to transform the JPA {@link Entity} instances
+   * @return the {@link ResourceTransformer} to use to transform the JPA {@link Entity} instances
    *     into FHIR {@link ClaimResponse} instances
    */
   public ResourceTransformer<Claim> getTransformer() {
